requirements.txt
python-dotenv
scrapy
beautifulsoup4
pandas
pymongo
sentence_transformers
streamlit
openai
flask
flask-cors
